Fixed create.sql, fixed Bidder Location Parsing

+++ create.sql
@@ -29,7 +29,7 @@ create table ItemLocation(
     LocId int);
 create table Location(
     LocId int primary key, 
-    lat float, 
-    lon float, 
+    lat float(10,8), 
+    lon float(10,8), 
     locText varchar(255), 
     country varchar(255));
